# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp AI SaaS API",
    "description": "Multi-tenant WhatsApp sessions, contacts, send-message, and Gemini AI agents (one agent per session). Use **Authorize** with a JWT from `/api/auth/login` or `/api/auth/register`. All successful JSON responses use `{ \"success\": true, \"data\": ... }` unless noted.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local (change in Swagger if needed)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste token only (Swagger adds Bearer prefix)"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      },
...
      "OkTrue": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "data": {
            "type": "object",
            "properties": { "ok": { "type": "boolean", "enum": [true] } }
          }
        }
      },
      "SyncContactsData": {
        "type": "object",
        "properties": {
          "upserted": { "type": "integer", "minimum": 0 }
        }
      },
      "ContactItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "MongoDB ObjectId (hex)" },
          "phoneNumber": { "type": "string" },
          "name": { "type": "string" },
          "isBusiness": { "type": "boolean" },
          "lastSeen": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ContactListData": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ContactItem" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" }
        }
      },
      "ContactListResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "data": { "$ref": "#/components/schemas/ContactListData" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/contacts/{sessionId}/sync": {
      "post": {
        "tags": ["Contacts"],
        "summary": "Sync contacts from WhatsApp snapshot",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "Session public UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": { "type": "boolean", "enum": [true] },
                    "data": { "$ref": "#/components/schemas/SyncContactsData" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "SESSION_OFFLINE",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/contacts/{sessionId}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "List contacts",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 200 }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ContactListResponse" }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "400": {
            "description": "Invalid query (limit/page)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/contacts/{contactId}": {
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contact",
        "description": "`contactId` is the MongoDB ObjectId (24 hex chars), not the session UUID.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^[a-fA-F0-9]{24}$" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OkTrue" }
              }
            }
          },
          "400": {
            "description": "INVALID_ID",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    }
  }
}